Backend/requirements/requirements.txt
# requirements.txt
# Django & DRF Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9

# Environment & Security
python-dotenv==1.0.0
django-environ==0.11.2

# Image Processing (for user profiles)
Pillow==10.0.0

# Development Tools
django-debug-toolbar==4.2.0

# AI Tools (commented out for now)
# langchain==0.0.329
# chromadb==0.4.14
# sentence-transformers==2.2.2
# python-jose[cryptography]==3.3.0
